Move project towards a more general mkElmDerivation function
This commit changes the readme to give an example of the new way to
use the mkElmDerivation function. It is more general now, and subsumes
the existing other functions that this repository provides. This
change will mean that the amount of code in this repository decreases
and it will be more maintainable.